--- full-stack-tests/native-app/src/backend/RpcImpl.ts
@@ -1,46 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-* See LICENSE.md in the project root for license terms and full copyright notice.
-*--------------------------------------------------------------------------------------------*/
-import * as nock from "nock";
-import { ClientRequestContext, Config } from "@bentley/bentleyjs-core";
-import { IModelBankClient } from "@bentley/imodelhub-client";
-import { IModelJsFs, NativeHost } from "@bentley/imodeljs-backend";
-import { RpcInterface, RpcManager } from "@bentley/imodeljs-common";
-import { AuthorizedClientRequestContext } from "@bentley/itwin-client";
-import { CloudEnvProps, TestRpcInterface } from "../common/RpcInterfaces";
-import { CloudEnv } from "./cloudEnv";
-
-export class TestRpcImpl extends RpcInterface implements TestRpcInterface {
-  public static register() {
-    RpcManager.registerImpl(TestRpcInterface, TestRpcImpl);
-  }
-
-  public async getCloudEnv(): Promise<CloudEnvProps> {
-    const requestContext = ClientRequestContext.current as AuthorizedClientRequestContext;
-    if (CloudEnv.cloudEnv.isIModelHub) {
-      const region = Config.App.get("imjs_buddi_resolve_url_using_region") || "0";
-      return { iModelHub: { region } };
-    }
-    const url = await (CloudEnv.cloudEnv.imodelClient as IModelBankClient).getUrl(requestContext);
-    return { iModelBank: { url } };
-  }
-
-  public async purgeStorageCache(): Promise<void> {
-    return IModelJsFs.purgeDirSync(NativeHost.appSettingsCacheDir);
-  }
-
-  public async beginOfflineScope(): Promise<void> {
-    nock(/^https:\/\/.*$/i)
-      .log((message: any, optionalParams: any[]) => {
-        // eslint-disable-next-line no-console
-        console.log(message, optionalParams);
-      }).get("/").reply(503);
-  }
-
-  public async endOfflineScope(): Promise<void> {
-    nock.cleanAll();
-  }
-
-}
-TestRpcImpl.register();
