+++ full-stack-tests/native-app/src/backend/backend.ts
@@ -5,19 +5,17 @@
 
 // Sets up certa to allow a method on the frontend to get an access token
 import "@bentley/oidc-signin-tool/lib/certa/certaBackend";
-import "./RpcImpl";
+import * as nock from "nock";
 import * as path from "path";
-import { BentleyLoggerCategory, Logger, LogLevel } from "@bentley/bentleyjs-core";
+import { BentleyLoggerCategory, ClientRequestContext, Config, Logger, LogLevel } from "@bentley/bentleyjs-core";
 import { loadEnv } from "@bentley/config-loader";
 import { ElectronHost } from "@bentley/electron-manager/lib/ElectronBackend";
-import { IModelHubClientLoggerCategory } from "@bentley/imodelhub-client";
-import { BackendLoggerCategory, IModelHostConfiguration, NativeLoggerCategory } from "@bentley/imodeljs-backend";
+import { IModelBankClient, IModelHubClientLoggerCategory } from "@bentley/imodelhub-client";
+import { BackendLoggerCategory, IModelHostConfiguration, IModelJsFs, IpcHandler, NativeHost, NativeLoggerCategory } from "@bentley/imodeljs-backend";
 import { RpcConfiguration } from "@bentley/imodeljs-common";
-import { ITwinClientLoggerCategory } from "@bentley/itwin-client";
-import { rpcInterfaces } from "../common/RpcInterfaces";
+import { AuthorizedClientRequestContext, ITwinClientLoggerCategory } from "@bentley/itwin-client";
+import { CloudEnvProps, testIpcChannel, TestIpcInterface } from "../common/IpcInterfaces";
 import { CloudEnv } from "./cloudEnv";
-
-/* eslint-disable no-console */
 
 function initDebugLogLevels(reset?: boolean) {
   Logger.setLevelDefault(reset ? LogLevel.Error : LogLevel.Warning);
@@ -38,6 +36,36 @@ export function setupDebugLogLevels() {
   initDebugLogLevels(false);
 }
 
+class TestIpcHandler extends IpcHandler implements TestIpcInterface {
+  public get channelName() { return testIpcChannel; }
+
+  public async getCloudEnv(): Promise<CloudEnvProps> {
+    const requestContext = ClientRequestContext.current as AuthorizedClientRequestContext;
+    if (CloudEnv.cloudEnv.isIModelHub) {
+      const region = Config.App.get("imjs_buddi_resolve_url_using_region") || "0";
+      return { iModelHub: { region } };
+    }
+    const url = await (CloudEnv.cloudEnv.imodelClient as IModelBankClient).getUrl(requestContext);
+    return { iModelBank: { url } };
+  }
+
+  public async purgeStorageCache(): Promise<void> {
+    return IModelJsFs.purgeDirSync(NativeHost.appSettingsCacheDir);
+  }
+
+  public async beginOfflineScope(): Promise<void> {
+    nock(/^ https: \/\/.*$/i)
+      .log((message: any, optionalParams: any[]) => {
+        // eslint-disable-next-line no-console
+        console.log(message, optionalParams);
+      }).get("/").reply(503);
+  }
+
+  public async endOfflineScope(): Promise<void> {
+    nock.cleanAll();
+  }
+}
+
 async function init() {
   loadEnv(path.join(__dirname, "..", "..", ".env"));
 
@@ -52,6 +80,6 @@ async function init() {
   iModelHost.concurrentQuery.concurrent = 2;
   iModelHost.concurrentQuery.pollInterval = 5;
   iModelHost.cacheDir = path.join(__dirname, "out");
-  await ElectronHost.startup({ electronHost: { rpcInterfaces }, iModelHost });
+  await ElectronHost.startup({ electronHost: { ipcHandlers: [TestIpcHandler] }, iModelHost });
 }
 module.exports = init();
