+++ full-stack-tests/native-app/src/frontend/_Setup.test.ts
@@ -3,31 +3,7 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 
-import { assert } from "chai";
-import { BentleyCloudRpcManager, RpcConfiguration } from "@bentley/imodeljs-common";
-// Testing order of initialization for imodeljs-frontend and imodeljs-common
-import { rpcInterfaces } from "../common/RpcInterfaces";
-import { ProcessDetector } from "@bentley/bentleyjs-core";
+import { RpcConfiguration } from "@bentley/imodeljs-common";
 
 RpcConfiguration.developmentMode = true;
 RpcConfiguration.disableRoutingValidation = true;
-
-if (!ProcessDetector.isElectronAppFrontend) {
-  const config = BentleyCloudRpcManager.initializeClient({ info: { title: "full-stack-test", version: "v1.0" } }, rpcInterfaces);
-  config.protocol.pathPrefix = `http://${window.location.hostname}:${Number(window.location.port) + 2000}`;
-
-  // This is a web-only test
-  describe("Web Test Fixture", () => {
-    it("Backend server should be accessible", async () => {
-      const req = new XMLHttpRequest();
-      req.open("GET", `${config.protocol.pathPrefix}/v3/swagger.json`);
-      const loaded = new Promise((resolve) => req.addEventListener("load", resolve));
-      req.send();
-      await loaded;
-      assert.equal(200, req.status);
-      const desc = JSON.parse(req.responseText);
-      assert.equal(desc.info.title, "full-stack-test");
-      assert.equal(desc.info.version, "v1.0");
-    });
-  });
-}
